Split function componentts for each page into their own file
