+++ src/Add.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Add() {
+	return (
+			<h1>Add an item to your list here</h1>
+	)
+}
+
+export default Add;
