+++ src/Home.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Home() {
+	return (
+			<h1>congrats this is the home page</h1>
+	)
+}
+
+export default Home
