+++ src/List.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function List() {
+	return (
+			<h1>This is the list page!</h1>
+	)
+}
+
+export default List;
